Lots of fixes to Component classes and LevelLoadCompleteHandler. Not quite world loading, but getting closer...

+++ ImaginationServer.Common/Data/BackpackItem.cs
@@ -3,7 +3,7 @@
     public class BackpackItem
     {
         public virtual int Id { get; set; }
-        public virtual long Lot { get; set; }
+        public virtual uint Lot { get; set; }
         public virtual int Slot { get; set; }
         public virtual int Count { get; set; }
         public virtual bool Linked { get; set; }
